Delay hiding hoverable tooltips
It's easy to overshoot the bottom of the tooltip when cursoring to a
button, such as opening the commit from a blame tooltip. Before this
change the tooltip would immediately disappear, and now it sticks
around for a bit.

Also:

* Shares the implementation with `elements/text.rs`. This will
particularly be handy when it makes use of hoverable tooltips. Also
means that it will use the same handling for mouse down / mouse
wheel (it did not handle mouse wheel before).

* Improves the fix to #21657.

    - Now the element will no longer think it has an active tooltip
    that it registers with the window.

    - It will instead display the next available tooltip, whereas
    I believe before the next available tooltip would be suppressed.

* Fixes bug where `cx.refresh()` wasn't called when text tooltip is
hidden due to a mouse down event.

* Ports over fix in #14832 to `elements/text.rs`